Adding Cache and Changing from Integer to Long

+++ API/src/main/java/org/acme/Alticci.java
@@ -2,6 +2,7 @@ package org.acme;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
+import java.math.*;
 
 
 @Path("/alticci")
@@ -11,7 +12,7 @@ public class Alticci {
     @Path("/{id}")
     public String alticci(String id) {
         //Transform the String into a integer
-        int value = Integer.parseInt(id);
+        long value = Long.parseLong(id);
         //The numbers can't be negative so the loop doesn't go infinite
         if (value < 0) return "Numero tem de ser maior ou igual a 0";
         //returning the function "Alticci"
